requirements.txt
Django==4.0.5
python-decouple==3.6
psycopg2-binary==2.9.3
gunicorn==20.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-colorfield==0.7.0
Pillow==9.1.1
django-filter==21.1
aiohttp==3.8.1
celery==5.2.7
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
gevent==21.12.0
django-allauth==0.51.0
dj-rest-auth==2.2.4
six==1.16.0
requests==2.27.1
beautifulsoup4==4.11.1
redis==4.3.3
websockets==10.3
channels==3.0.4
daphne==3.0.2
channels-redis==3.4.0
PyJWT==2.4.0
asgiref==3.5.2
drf_yasg==1.20.0
qrcode==7.3.1
arabic-reshaper==2.1.3
python-bidi==0.4.2